+++ src/main/java/de/telekom/sea3/spring/model/Size.java
@@ -1,17 +1,17 @@
 package de.telekom.sea3.spring.model;
 
 public class Size {
-    private int size;
-    public Size(int size){
+    private Long size;
+    public Size(Long size){
         super();
         this.size=size;
     }
 
-    public int getSize() {
+    public Long getSize() {
         return size;
     }
 
-    public void setSize(int size) {
+    public void setSize(Long size) {
         this.size = size;
     }
 }
